jhbuild: Add moduleset with tarballs used for 7.6 katamari
Add a moduleset for the 7.6 katamari based on xorg.modules and
module-list.txt. This uses the module list from 7.6 with gzipped
tarballs and sha1sums from announce emails (when available).

Modules not included in the katamari have been given their most recent
release. It might be a good idea to group typical addons such as xinit
in an "extras" metamodule.

Originally some modules had autogen-sh="configure" specified in their
definitions. This forced configure to be used even if autogen.sh was
present. This has been fixed in jhbuild commit ff2211e6.

Signed-off-by: Dan Nicholson <dbn.lists@gmail.com>