build.sh: provide a new implementation for the -g option
Given that CFLAGS variable is now supported, the -g option
is reimplemented using CFLAGS. The debug options are appended
to CFLAGS, preserving the user's existing options, if any.

The CONFCFLAGS variable is removed. It had not been designed
as a vehicle to forward C flags to module configuration, but
as a -g implementation. A statement such as
export CONFCFLAGS=-"g3 -O0" would fail.

Reviewed by: Trevor Woerner <twoerner@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>