build.sh: add support for EPREFIX for architecture-dependent files
Follow Autoconf file architecture as there is a PREFIX for
architecture-independent files (config, docs, man) and an EPREFIX
for architecture-dependent files (bin, lib, exe).

This summarizes the architecture-dependent files directories:
EPREFIX:
  --bindir=DIR            user executables [EPREFIX/bin]
  --sbindir=DIR           system admin executables [EPREFIX/sbin]
  --libexecdir=DIR        program executables [EPREFIX/libexec]
  --libdir=DIR            object code libraries [EPREFIX/lib]

This summarizes the architecture-independent files directories:
PREFIX:
  --sysconfdir=DIR        read-only single-machine data [PREFIX/etc]
  --sharedstatedir=DIR    modifiable architecture-independent data [PREFIX/com]
  --localstatedir=DIR     modifiable single-machine data [PREFIX/var]
  --includedir=DIR        C header files [PREFIX/include]
  --datarootdir=DIR       read-only arch.-independent data root [PREFIX/share]

The default value for EPREFIX is PREFIX as explained in the help:
  --prefix=PREFIX         install architecture-independent files in PREFIX
                          [/usr/local]
  --exec-prefix=EPREFIX   install architecture-dependent files in EPREFIX
                          [PREFIX]

The build.sh script made it impossible to use a value for EPREFIX other
than the default value.

The patch adds the EPREFIX env variable and repleaces PREFIX with EPREFIX
where appropriate. If a value other than the default value is used,
the CONFCMD emitted will reflect that with --exec-prefix option.

Reviewed-by: Trevor Woerner <twoerner@gmail.com>
Reviewed-by: Matt Dew <matt@osource.org>
Acked-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>