build.sh: allow user to specify an alternate bin directory
Any module can be configured with --bindir but the way the PATH
is setup by build.sh does not allow this freedom. It asumes
that the bin subdir will always be under $EPREFIX.

The default value is "bin" under $EPREFIX but that does not mean
that any alternate subdir value must be under $EPREFIX.
The prefix might be /usr but the executables can be under $HOME/bin.

A BINDIR env variable is added which works in a similar way to
PREFIX and EPREFIX. It follows the Autoconf convention.

If a value other than the default value is used,
the CONFCMD emitted will reflect that with --bindir option.

Reviewed-by: Trevor Woerner <twoerner@gmail.com>
Reviewed-by: Matt Dew <matt@osource.org>
Acked-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>