build.sh: allow user to specify an alternate datadir
Modules can be configured with --datadir to provide an alternate
location for architecture-independent data.

The ACLOCAL dir was created relative to the hard-coded "share" subdir.
Same for PKG_CONFIG_PATH.

A DATAROOTDIR variable is added which provides the "share" default
value to all directories based on it, including DATADIR

A DATADIR variable is added which has DATAROOTDIR as a default value.

If a value other than the default value is used,
the CONFCMD emitted will reflect that with --datadir or -datarootdir option.

There are other directories based on datarootdir but they are not needed
to setup the build environment. Values for these directories can be
passed using CONFFLAGS env variable.

Reviewed-by: Trevor Woerner <twoerner@gmail.com>
Reviewed-by: Matt Dew <matt@osource.org>
Acked-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>