build.sh: allow user to specify an alternate localstatedir
Modules can be configured with --localstatedir to provide an alternate
location for modifiable single-machine data.

The log dir was created relative to the hard-coded "var" subdir.

A LOCALSTATEDIR variable is added which uses the "var" default value.

If a value other than the default value is used,
the CONFCMD emitted will reflect that with --localstatedir option.

Reviewed-by: Trevor Woerner <twoerner@gmail.com>
Reviewed-by: Matt Dew <matt@osource.org>
Acked-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>