build.sh: allow user to specify an alternate libdir
Modules can be configured with --libdir to provide an alternate
location for object cod libraries. The default value is EPREFIX/lib.

The PKG_CONFIG_PATH, LD_LIBRARY_PATH and FONTPATH are set with
a value of PREFIX/$LIBDIR which does not allow a different prefix
as does the --libdir option.

The meaning of the exisitng LIBDIR is aligned with similar variables
to allow the user to provide a libdir that is not restricted to be
under PREFIX.

If a value other than the default value is used,
the CONFCMD emitted will reflect that with --libdir option.

For environments already having LIBDIR set as a path segment
(e.g. "lib64"), this will no longer work. The next patch will perform
error checking, ensuring all these *DIR variables are full paths

This will also fix the annoyance where the --libdir was always
emitted even if the default value was used.

Reviewed-by: Trevor Woerner <twoerner@gmail.com>
Reviewed-by: Matt Dew <matt@osource.org>
Acked-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>