build.sh: do not set or document the xserver FONTPATH env variable
This variable was introduced in 2005 at a time where the xserver did
not supply a default value for FONTPATH. Now it does, and it is more
complex that this script handle.

If the variable is set in the build environment by the user, it will continue
to be picked-up by the xserver.

If the variable was set by the script, its incorrect value will no longer
be used by the xserver as it provides a correct default value.

Reviewed-by: Trevor Woerner <twoerner@gmail.com>
Reviewed-by: Matt Dew <matt@osource.org>
Acked-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>