build.sh: make PREFIX an optional variable as in Autoconf configure
The prefix is optional when configuring a package with autoconf.
The script assumes a value of /usr/local if missing, as configure does.
This makes the usage between build.sh and configure more consistent.

Setting PREFIX in the environment is an alternative to setting prefix
on the command. A reminder that the prefix is not locate the source code
but to specify the base installation directory for binaries, docs, etc...

Reviewed-by: Trevor Woerner <twoerner@gmail.com>
Reviewed-by: Matt Dew <matt@osource.org>
Acked-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>