build.sh: rebuild a failing target with Automake silent rules disabled
With Automake silent rules enabled, the compiler error message is
displayed but not the command that caused it.

This patch will re-issue the make command that failed, but this time
with the V=1 alongside the MAKEFLAGS env variable. A message is issued
to not confuse the user as to why the target failed twice.

This introduces a new cmd line option --retry-v1. The "v1" part is a hint
to those familiar with Automake silent rule variable "V".
There is no backing environment variable for this option.

Reported-by: Pat Kane <pekane52@gmail.com>

Tested-by: Trevor Woerner <twoerner@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>