jhbuild: Build openchrome driver from svn
Since the openchrome driver is being distributed in the katamari, it
makes sense to build it here even if it's not under X.Org.

The openchrome svn setup is a little unusual. Rather than having the
code in it's own project subdirectory, it's directly in the root "svn"
directory. If "svn" is used as the branch module, jhbuild will use the
trunk or branches directories as directed by the revision attribute.

Signed-off-by: Dan Nicholson <dbn.lists@gmail.com>
Reviewed-by: Dirk Wallenstein <halsmit@t-online.de>