build.sh: better integrate --autoresume and -n
The --autoresume <file> option allows a user to specify a <file> into which
the build prints each module/component it has built. When a subsequent build
is restarted with <file>, the build can skip all previously built modules,
start with the last one (which is assumed to have failed previously), and
continue on.

The -n option allows a build to continue with subsequent modules even if
one or more of the modules fails to build correctly.

With this change, in addition to updating the --autoresume <file> with the
name of the module/component just built, <file> is also updated with the
status of the build. Therefore if you use -n you will have an --autoresume
<file> which lists the build status of all the modules you wanted to build.

A subsequent build using the --autoresume <file> will scan <file> looking
for failures and attempt to build them before continuing on with the build
from the end of the list.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>