build.sh: make --autoresume and -o mutually exclusive
The '--autoresume' option tells the build to continue on with where it last
left off while building a set of modules. The '-o' option tells the build to
build just the one module. These two options don't make a lot of sense when
specified together. This patch looks for the scenario where both these
options are specified and errors out since there is a good chance the build
will not behave as the user expects.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>