build.sh: no longer require a trailing slash for module name
A few modules have a single level directory such as pixman, xserver,
and so on. Currently the script requires specifying such modules in
the module list file or with the -o option with a trailing slash. So one needs
to learn and remember using pixman/ or xserver/.

The patch will correct the situation such that both input are accepted.
The generated list (-L option) will write the module names without the
trailing slash.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>