build.sh: tarballs not found under their respective module directory
Currently, the module name is changed to add a subdir. For example "xcb" is
changed to "xcb/proto". The tarball cannot be found under xcb as expected.
This is done for pixman, xcb and mesa only, not for app, driver, lib, or any
other module which is inconsistent.

Moving the tarball under xcb/proto works, but yields a build structure
that is unpractical: xcb/proto/xcb-proto-1.8. The source code is now
under two subdirs rather than the usual one.

The package archive already contains the component directory, suffixed
with a version number.

By not changing the module name, we have xcb/xcb-proto-1.8. The same
pattern holds true for mesa and pixman. Changing the module name has
other adverse effects in subsequent functionality.

Note that for pixman, this module is under the "pixman" directory
and is not a component of the "lib" module. The module directory
structure is given by git and not by the location of the tarballs
on the web.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>