build.sh: extract tarballs in a consistent location
The tarballs are extracted in the same location where they are found. So libXi
code will be extracted in the toplevel directory if this is where the tarball
is found. Then lib/libX11 code will be extracted under lib subdir if, for
example, the tarball is placed in the lib subdir.

This patch promotes a consistent build structure based on what we see when
working in git. The source code for modules having components (e.g. app,
driver) is extracted under the module directory. For modules with no
components (e.g. xserver), the tar file is extracted under the toplevel
directory. This is done regardless of where the tarball is found, either
in the toplevel directory, or under the module directory.

Sample build directory hierarchy:
app/luit-1.1.1
driver/xf86-input-void-1.4.0
font/font-util-1.3.0
lib/libXi-1.7
mesa/libdrm-2.4.47
mesa/Mesa-9.2.1
pixman-0.30.2
xcb/libxcb-1.9
xcb/xcb-proto-1.8
xkeyboard-config-2.5
xorg-server-1.14.99.2

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>