jhbuild: add missing modules and meta modules
The X Window System JHBuild Modules List Sample File

The list is intended to be exhaustive and covers all X modules and their
dependencies, roughly matching the latest release.

With the matching jhbuildrc in $HOME/.config, run 'jhbuild build'
to build all of X. The dependencies are such that you never need to
type the name of a meta module.

It is a sample, so you will need to copy and trim it to suit your needs.
Not all modules build on all O/S and not on all CPUs. The build order
is subject to dependencies, it will not follow the list order.

Most modules will build on platforms they are not intended to run on.
This allows working on Autoconf, man pages, documentation, .pc files,
creating a tarball, and so on...

Refer to this excellent short guide on how to invoke JHBuild commands:
https://developer.gnome.org/jhbuild/unstable/index.html.en

The meta-modules are at the top of the file so it is easy to customize.

Some modules will be relocated on disk as they now follow the git URL.
Some modules have been renamed (e.g. mesa) so they are easier to find
for the non-expert. Still, inconsistencies remain.

The meta modules for broken drivers or platform specific drivers
are gone. Users are expected to copy and trim the file. There is
no "preferred" platform.

Keep in mind this is a *sample* that needs to be adapted to your environment.
It is much easier to remove than to add.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>