release.sh: implement around git worktree
Months ago, before my work in here there the script had a number of
assumptions:
 - autoreconf or alike is ran by the user - slight nuisance esp. when
releasing multiple packages.
 - the generated files were compatible with the ones on the server -
rarely and issue, but still
 - config.status means 'all the autotools bits were generated correctly'
 - failed to consider if user has multiple config.status - using
./build-64bit/ and ./build-32bit/ anyone ?
 - ...

With 663364cda5e316a0509ff5869293e3a815b9945f we mitigated a lot of that
but did not consider
 - forcing people to `git clean' is annoying and if not careful one can
purge local files that they want to keep
 - the files generated by autotools were left behind

In order to mitigate those use git worktree. This is more efficient
[both bandwidth and storage wise] than pulling/generating git tarballs,
git clone and friends.

Cc: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>