release.sh: Make project slug detection a bit more robust
Typical gitlab remote urls look like:

    git@gitlab.freedesktop.org:xorg/driver/xf86-video-v4l.git

The existing code would fail to parse this, because it looks for /xorg/
not :xorg/. So first, fix the match pattern. After that we'd break down
the slug wrong because we're counting /s wrong. Rather than carefully
know how many to care about: reverse the string, pick out the first two
fields from that, and reverse the result, so we're always picking off
the last two fields of the URL whatever it is.

Signed-off-by: Adam Jackson <ajax@redhat.com>